2017-07-19 Node.js Version 8.2.0 (Current) Release
Git-EVTag-v0-SHA512: e2083a0620180f81ac225471151aadb710d73c6289de0503123ebfc7c5fc0f454d62255ca1590c931bae11184bf3120353f0dc0837dbeeac21d0bfb9f4d184fa

+++ src/node_file.cc
@@ -494,10 +494,13 @@ void FillStatsArray(v8::Local<v8::Float64Array> fields_array,
 #else
   fields[9] = -1;
 #endif
-  // Dates.
-#define X(idx, name)                          \
-  fields[idx] = (s->st_##name.tv_sec * 1e3) + \
-                (s->st_##name.tv_nsec / 1e6); \
+// Dates.
+// NO-LINT because the fields are 'long' and we just want to cast to `unsigned`
+#define X(idx, name)                                           \
+  /* NOLINTNEXTLINE(runtime/int) */                            \
+  fields[idx] = ((unsigned long)(s->st_##name.tv_sec) * 1e3) + \
+  /* NOLINTNEXTLINE(runtime/int) */                            \
+                ((unsigned long)(s->st_##name.tv_nsec) / 1e6); \
 
   X(10, atim)
   X(11, mtim)
